refactor: AuthInfo를 Credential로 재명명
